split up tests, so features have related testclasses
